Read tensorQTL results into R

Reads the output files written by a tensorQTL run into GenomicRanges::GRanges objects, matched to the features in the originating tQTLExperiment.

readTQTL(
  outDir,
  mode = c("cis_nominal", "cis", "cis_independent", "trans"),
  x = NULL
)

Arguments

outDir

The same directory passed to prepareTQTL().

mode

The mode used for the tensorQTL run.

x

The tQTLExperiment used to generate the input files, used to attach feature coordinates to region-level results.

Value

A named list of GenomicRanges::GRanges:

pairs

(cis_nominal) Feature-variant pairs with association statistics.

hits

(cis) One range per feature with permutation p-value.

output_files

(other modes) Paths to the raw output files.

See also

Examples

exdir <- system.file("extdata", package = "tQTLExperiment")
tqe <- tQTLExperiment(
    plinkPrefix = file.path(exdir, "chr22-n100"),
    phenoFile   = file.path(exdir, "mean-pheno-n100.bed")
)
#> Extracting number of samples and rownames from chr22-n100.fam...
#> Extracting number of variants and colnames from chr22-n100.bim...
od <- tempdir()
cmd <- prepareTQTL(tqe, outDir = od, mode = "cis_nominal")
#> Run the following command in a terminal with tensorqtl available:
#> 
#> python3 -m tensorqtl /private/var/folders/yw/gfhgh7k565v9w83x_k764wbc0000gp/T/RtmpbYmUMP/temp_libpath108bb1bea0898/tQTLExperiment/extdata/chr22-n100 /var/folders/yw/gfhgh7k565v9w83x_k764wbc0000gp/T//Rtmp6cglZJ/pheno.bed /var/folders/yw/gfhgh7k565v9w83x_k764wbc0000gp/T//Rtmp6cglZJ/tqtl_out --mode cis_nominal --maf_threshold 0.05 --window 1000000 -o /var/folders/yw/gfhgh7k565v9w83x_k764wbc0000gp/T//Rtmp6cglZJ
#> 
#> Then call readTQTL('/var/folders/yw/gfhgh7k565v9w83x_k764wbc0000gp/T//Rtmp6cglZJ', mode = 'cis_nominal') to load results into R.
# (user runs cmd in terminal)
# res <- readTQTL(od, mode = "cis_nominal", x = tqe)